fix(builder-main): Push to API canvas on element add
This commit moves the original `BuilderMain` code to a new file: `BuilderMainOld.vue` in order to resolve some issues when syncing draggable components and API canvas.

The `BuilderMain` file includes the following improvements:
- Remove of the old code and start fresh.
- Include base structure and draggable-component.
- Include `UIDevices` component.
- Fetch and sync API canvas.
- Include a method that will trigger when user adds a new element to the drop zone. This method will update the droppedElements and the API canvas.
